mybackend/requirements.txt
asgiref==3.4.1
certifi==2023.7.22
charset-normalizer==3.3.2
cloudinary==1.36.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.10
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.0
django-import-export==3.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg2==1.19.4
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
midtransclient==1.4.2
numpy==1.26.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==6.0.1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.4.2
tablib==3.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
